+++ test/java/io/github/michaelmior/automaton/SpecialOperationsTest.java
@@ -1,0 +1,56 @@
+/*
+ * io.github.michaelmior.automaton
+ *
+ * Copyright (c) 2001-2017 Anders Moeller,
+                 2025 Michael Mior
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. The name of the author may not be used to endorse or promote products
+ *    derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package io.github.michaelmior.automaton;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.util.HashSet;
+import java.util.Set;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.provider.MethodSource;
+
+/** Tests {@link Automaton} */
+final class SpecialOperationsTest {
+  @MethodSource("mergeStates")
+  @Test
+  void merges_states() {
+    Automaton a = new RegExp("abc").toAutomaton();
+
+    // Merge the first two states
+    Set<State> toMerge = new HashSet<>();
+    toMerge.add(a.getInitialState());
+    for (Transition t : a.getInitialState().getTransitions()) {
+      toMerge.add(t.to);
+    }
+
+    Automaton a2 = SpecialOperations.mergeStates(a, toMerge);
+    assertEquals("bc", a2.getShortestExample(true));
+  }
+}